+++ ChangeLog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.28 - unreleased]
+## [0.28]
 ### Added
 - compartmentalize backend property implementation into `PropertyKeyingBase`, a backend-specific decorator
 - implement `property.visa_keying` and `property.message_keying`